add button to revert to frequency base trip. move chart component and change the trash to be more clickabnle in the list of Ttrips
